[#20250701] Overhaul verbose mode
